chore: move request message creation to `network` pkg
